backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
python-multipart==0.0.6
pydantic[email]==2.5.0
pydantic-settings==2.1.0
alembic==1.12.1
python-dotenv==1.0.0
httpx==0.25.1
email-validator==2.1.0
jinja2==3.1.2
asyncpg<0.29.0
minio==7.2.7